+++ Entites/StockProfile/InsiderRoster.cs
@@ -3,6 +3,6 @@
 public class InsiderRoster
 {
     public string EntityName { get; set; }
-    public double? Position { get; set; }
+    public long? Position { get; set; }
     public long? ReportDate { get; set; }
 }
